refactor: Line 에서 getStations 를 Sections 에서 처리하도록 이동
